new file:   README.md 	deleted:    conf/modules.d/network.conf 	deleted:    conf/monigraf.conf 	deleted:    modules/network 	deleted:    monigraf.py 	renamed:    conf/modules.d/apcups.conf -> monigraf/conf/modules.d/apcups.conf 	new file:   monigraf/conf/modules.d/network.conf 	renamed:    conf/modules.d/smart.conf -> monigraf/conf/modules.d/smart.conf 	new file:   monigraf/conf/monigraf.ini 	renamed:    install.sh -> monigraf/install.sh 	renamed:    modules/apcups -> monigraf/modules/apcups 	new file:   monigraf/modules/network 	renamed:    modules/smart -> monigraf/modules/smart 	new file:   monigraf/monigraf.py 	renamed:    monigraf.service -> monigraf/monigraf.service 	deleted:    requirement 	new file:   requirements.txt
