--- conf/modules.d/smart.conf
@@ -1,2 +1,0 @@
-# Run this module every x seconds
-delay=3600
