--- conf/monigraf.conf
@@ -1,6 +1,0 @@
-# Influxdb Configuration
-influxserver="SERVER IP"
-userdb="USERNAME"
-password='PASSWORD'
-dbname="monigraf"
-
