+++ monigraf/conf/modules.d/network.conf
@@ -1,0 +1,5 @@
+# Run on these hosts
+hosts=("sevpn")
+
+# Run this module every x seconds
+delay=1800
