+++ monigraf/conf/modules.d/smart.conf
