+++ monigraf/conf/monigraf.ini
@@ -1,0 +1,20 @@
+# See https://github.com/JackScripter/monigraf for more information.
+[DEFAULT]
+MOD_PATH = /opt/monigraf/modules/
+MOD_ENABLED = ["network","smart"]
+delay = 3600
+
+[influxdb]
+server = SERVER IP
+username = monigraf
+password = 'PASSWORD'
+dbname = monigraf
+
+# Modules configuration
+[network]
+interfaces = all
+
+[apcups]
+
+[smart]
+drives = all
