+++ monigraf/modules/smart
