--- requirement
@@ -1,3 +1,0 @@
-influxdb-client
-smartmontools
-
