+++ requirements.txt
@@ -1,0 +1,1 @@
+paramiko==2.7.1
